+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 0.7.4
 ### Added
 - Bing Maps Support
+- [Iconclass](http://iconclass.org)-Attribute
 ### Fixed
 - Accidentally moving map outside viewport
 - Ignore uppercase/lowercase letters in login form user names